Created AI Gatherer
Created AI Gatherer class. This class will be used for AI that gather resources. They will go to the nearest resource node, get the resource and return it to a stockpile building. An AI interface has also been created for common functions. AI Unit has also had a lot of code removed as it was unnecessary.