Unit Move Order
Units can now be moved by selecting them with left click and selecting a new destination with a right click. The pathfinding will make them go to that location. The original entity based version of this has also been removed as it was not practical for use with the camera system.