Selectable Build System
Selection options for building have been created. This has temporarily been achieved by using the numbers on the keyboard. There is now also a gold cost for building each type of object and a model for each object. The Game resources now also has a function for removing gold. The RTS Control System has also been renamed back to the RTS build system as the order system will be in a seperate script for code management. A folder has also been created within the scripts folder to house the controllers, e.g. the Game Handler class.