Prototype House Model
A very basic model of a house has been added to the game. The prefabs now also have a child transform where this model is placed and any scripts will be placed upon the actual prefab, as opposed to being with the visual side of things. Other prefabs have been removed and will be reimplemented as their models are added.