Extended UI Animator
Extended the UI Animator to allow for different animations for enabling and disabling of the UI element. This makes the system very useful. Also added a setting on the tab group that disables the default behaviour of instantly disabling the tab to allow for this animation system to be used.