Fixed #2
Fixed the tab system glitching with the animation system. This was achieved by changing the method the tabs used to interact with the animation system. Instead of having the tabs enable the pages, they now manually can call the method to start the animation using the callbacks configured on the tab buttons. The current tween is also now cancelled if another is called on the same object. Also made it clear on the tab group script that the status refers to the colour. Also added ability to reselect the currently selected tab to refresh it if need be. This only occurs when the boolean is set on the tab group.