Disable Tab System Function
Added function to disable / reset the tab system.  This is used when the user presses escape to go back to the startup phase of the main menu. Made input to change menu state in the main menu manager better by preventing the user entering the same state they are already in and causing the UI to glitch.