UI Animator Improvements
Made use of scriptable objects to create a better way of managing the UI Animator system for the user. The user can now create an asset with all the correct settings. This asset can then be dragged into the UI Animator script. This saves time and causes less problems for when multiple of the same animation are required, for example in the tab based menus.