Resolved #3
Resolved bug. The solution to this is related to layout groups. The parent of the UI element should not have a layout group of any kind and the anchor mode of the object attached to it should be on stretched mode for best results. Corrected null pointer checking in UI Animator also.