Improved Button Default Behaviours
Improved the default behaviours for the button system. It now tries to find a UI animator on the game object instead of just enabling it. This means animations do not have to be manually added through the button callback, and can now just be implemented easily. Also changed the tab system animation page to use a different duration and ease type that fits better with the style.