Basic Scene Loading
Added basic scene loading. This is extremely basic and resembles a prototype. There are many unpolished features so far, including the lack of functionality on the loading bar and the lack of a background image. The loading screen also starts immediately after the button is pressed, not giving sufficient time for the UI animation to finish.