Loading Screen Transition Polishing
Made the loading screen transitions better when loading into the actual game scene. This was done using as simple UI Animation Asset for a loading screen fade. Also added functionality for the loading bar, not that it can be seen as the scene is not intense enough for the loading bar to be used as the scene loads instantly. Also added a null pointer check in the UI animator for if there is no parent to the animated object. A transition still needs to be done for the start of the loading screen but this will require the implementation of some sort of timing script or wait for an event such as when the tab system page animation finishes, the game begins loading.